Implement the foundation of Keymail (messaging in Peergos)
This creates the broadcast structure for a chat. A causally consistent CRDT of messages.

The model is each member maintains an append only file of messages in the chat,
which is shared with other members.
To send a new message they simply append it to their own copy with an appropriate TreeClock update.
Clients can choose to merge in messages from others at will.

This needs a generalization of a vector clock (TreeClock) to handle adding members.