Implement removal from chat
Any member can remove themselves, or an admin can remove anyone

Admins are a chat group property "admin" which is a comma separated list of usernames.

The creator of a chat is the original admin. Only admins can update the admin property.
Concurrent modifications are tie broken by admin id.

When removing another member, first the keys on the shared chat state and message log
are rotated. Then a message is sent telling others to do the same.